dev-requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=dev-requirements.txt dev-requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via
    #   httpx
    #   starlette
    #   watchfiles
autopep8==2.3.1
    # via -r requirements.in
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
dnspython==2.7.0
    # via
    #   email-validator
    #   pymongo
ecdsa==0.19.0
    # via python-jose
email-validator==2.2.0
    # via fastapi
faker==30.6.0
    # via -r dev-requirements.in
fastapi[standard]==0.115.2
    # via -r requirements.in
fastapi-cli[standard]==0.0.5
    # via fastapi
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.6
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.27.2
    # via fastapi
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
jinja2==3.1.4
    # via fastapi
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycodestyle==2.12.1
    # via autopep8
pydantic==2.9.2
    # via fastapi
pydantic-core==2.23.4
    # via pydantic
pygments==2.18.0
    # via rich
pymongo==4.10.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via faker
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   uvicorn
python-jose==3.3.0
    # via -r requirements.in
python-multipart==0.0.12
    # via fastapi
pyyaml==6.0.2
    # via uvicorn
rich==13.9.2
    # via typer
rsa==4.9
    # via python-jose
shellingham==1.5.4
    # via typer
six==1.16.0
    # via
    #   ecdsa
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.40.0
    # via fastapi
typer==0.12.5
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   faker
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typer
uvicorn[standard]==0.32.0
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-cli
watchfiles==0.24.0
    # via uvicorn
websockets==13.1
    # via uvicorn
